Add AWK based commit message validator in shell
This version of the commit message validator uses AWK and
able to print the line number where the problem is found.
Points where is the problem within the given line.

It has a unit test to make sure the functionality is working fine.

Script has been created to add the check to the commit-msg hook.

Signed-off-by: István Rejtõ <reistvan@gmail.com>