Fix punctuation handling
There was a bug in the code which caused the script to reject commit
message when it ended with uppercase letter. The cause of the problem
was the unescaped '-' (dash) character in the pattern.
Awk recognized as an interval from ';' to '_'. This was fixed and
test were added to verify the result.

Fix: Issue #1

Signed-off-by: István Rejtõ <reistvan@gmail.com>